Some changes
Whether a material has food (apples and carrots) is now a boolean in the config. 1 in 4 chance for random materials to have food.
Whether a material has horse armour is now dependent on the regular armour boolean.
Armour now has only a 1 in 4 chance to have a toughness value.
Food is now in its own item group.
Added a new utility function for getting a boolean with a 1 in (x) chance to be true.